[workload] Fix using empty RuntimeIdentifier in LibraryBuilder AutoImport.props
When `mobile-librarybuilder-experimental` is installed, the LibraryBuilder SDK AutoImport.props activates and it incorrectly tries to IndexOf RuntimeIdentifier even when it's an empty string. This can lead to annoying error messages popping up even when not using the workload.

The fix is to add a condition only when RuntimeIdentifier has a value.