Change the stresslog message layout to support extra-large modules and modules loaded at far-apart addresses that use the stresslog.
Today, stresslog message strings that are more than 2^25 bytes after the start of the native module that contains the Stresslog code are recorded correctly. This leads to a few problems:

- If CoreCLR is too big (for example if it is instrumented by native sanitizers), the strings section can be too far away because of the extra code.
- If a user requests we load an out-of-module GC, it could be placed too far away.
- In NativeAOT, we don't control where the strings are placed and the runtime is linked into the same binary as the user's managed code. This can result in strings being placed out of range in large NativeAOT apps.

This change allows us to use format strings at any offset, at the cost of each message using more memory (so less messages recorded for the same allocated space).