HACK: kind of work around the emscripten_runtime_keepalive_push/pop no-op
Keep a thread local counter in mono. that we "know" will be right for
us and manually call unwind_to_js and thread_exit.

This is super-fragile since we don't know what emscripten internals
might be trying to manipulate the keepalive count and also we are
exiting the thread with active managed frames, so we might be skipping
finally clauses and possibly leaking interpreter memory.

This is mainly meant to keep work going on this branch and not
something we necessarily want to commit