Coalesce identical constant WordCategories
There are four places the same mask is computed in this project, can't eliminate the one in RegexGenerator.Emitter or its matching unit test because that's outputting the actual C# source string.
Changed the name `WordCategoriesMask` to avoid a conflict with  existing member.
Removed the detritus in the comments from the copy-paste from Emitter source (trailing `"` and `,`)