Remove the new node, it's likely not needed.
The only case where it might matter is if there's a handle access to a GVM. Then we go through this code:
https://github.com/dotnet/runtime/blob/57ae91cf6fc5672e34705b1a272cf268761d505a/src/coreclr/tools/aot/ILCompiler.Compiler/Compiler/DependencyAnalysis/RuntimeMethodHandleNode.cs#L60

That will potentially call the `GetDependenciesDueToMethodCodePresence` multiple times on the same method.

But that should not affect static cctors (since they're not GVMs) so the `GetDependenciesDueToMethodCodePresense` should only be called once per static cctor.

And even if it does happen (in the future) all it would cause is potentially generating duplicate warnings, no real functional problems.