Add wrappers only AOT option.
Library mode has a use case to support UnmanagedCallersOnly
(native-to-managed) and direct pinvokes (managed-to-native) and
still fallback to JIT for majority of managed methods. This commit
adds an undocumented AOT option to only generate wrappers and when
running that option together with partial AOT, only
native-to-managed and managed-to-native wrappers are generated.