Set better DfsBlockEntry stack initial capacity
The stack required by the DFS traversal can vary widely based on
the characteristics of the flow graph. However, it tends to increase
with the number of basic blocks in the function. To avoid excessive
ArrayStack reallocation, set an initial that is 60% of the number
of BasicBlocks in the function, with a minimum of the ArrayStack
built-in size. Round up to a power of two, and cap at 64 to avoid
excessive initial allocation with outlier functions.

With benchmarks MCH, this reduces reallocations from 39674 to 12858
and total stack slots allocated from 962480 to 908880. For
libraries.crossgen2 MCH, reallocations reduce from 33528 to 8498 and
total stack slots allocated from 777280 to 736368.

Various global inline bit operation utility functions were moved
from compiler.hpp to utils.h, so they can be used more broadly. They
were not related to Compiler.