Re-implement reachability
Currently, the JIT reachability sets implementation walks over every
block in the function computing the reaching sets using the reaching
set of each predecessor of the block. If there are any changes, every
block is iterated again. It walks the blocks in reverse postorder to
try and minimize the number of iterations required.

Re-implement this with a worklist. Seed the worklist with all the
"entry blocks" (the first block and all EH filters/handlers). For each
block, compute reachability. If the reachability changed, add all the
block's successors to the worklist. The worklist is ordered by reverse
post-order number, to try and reduce iterations.

This process reduces the total number of blocks processed (so improves
convergence). For the benchmarks collection (win-x64), it is reduced from
2,149,057 to 1,329,671 blocks processed. However, the successors processing
is expensive, and we also need to 'or' in the block itself every time
(as part of proper initialization and change determination).