Fix incorrect covariant returns tests
The intention of the methods is to override a base method, but the signature is wrong so they introduce a new slot instead. The test is not testing what it wanted to tests (that the override is not reachable) because we would never call this new slot method.

`!0` and `!1` corresponds to `UNUSED1` and `UNUSED2`.