Unconditionally set CLRTestExpectedExitCode.
Otherwise, a merged test group's wrapper's exit code will override
any out-of-proc test in the group.

This should be safe because other places that set CLRTestExpectedExitCode are either
- Later in the file so will continue to override this
- In src\tests\baseservices\TieredCompilation\RunBasicTestWithMcj.cmd/sh
  but setting/using it locally

This will break setting it in the environment and overriding all occurrences, but
such usage wouldn't have the expected per-test expected exit values.