Delete support for cloned types
Cloned types are a concept from old Redhawk that was used to deduplicate generics in multiple DLLs at runtime. I think.

It was used in a limited fashion in .NET Native to support string literals - so that we can have frozen objects with a direct reference to the System.String MethodTable in both SharedLibrary.dll and the App.exe. The one in App.exe was a System.String MethodTable that had the Cloned bit set.

We don't use this in Native AOT and a shared library is unlikely to come back. I need a bit to represent function pointer MethodTables and I intend to reuse this bit.