JIT: fix exit edge likelihood computation for capped loops
The initial version of this only really worked for very simple cases.
Generalize to handle loops with multiple exits better.

The rough idea is to compute how much additional profile flow needs
to come out of the loop to match the capped probability, and then
try and find one exit block where we can adjust likelhoods to cause
exactly this amount of extra flow to leave the loop.

Closes #84789.