Make `MethodTable` base size access more "typed"
We reuse the base size field for other purposes for `MethodTable`s that cannot actually exist on heap (otherwise GC needs this to be legit). I almost reused it for similar reasons for function pointers, but looks like we don't need it for those, so just submitting these asserts that we don't read the wrong meaning of the field.