Split PortableThreadPool.WorkerThread start and loop body (#84490)
This is Part 1 of https://github.com/dotnet/runtime/issues/84489 - landing support for async JS interop on threadpool threads in multi-threaded WebAssembly.

We will need to start the threadpool worker threads on the browser in a special way, such that they can exit back to the JS event loop, and use callbacks to run the worker loop body.

The current PR splits into a separate file the logic for starting threadpool worker threads, and the outer loop that waits for the semaphore that signals that work is available for the worker. The loop body (to be shared with the callback-based approach in a future PR) remains in PortableThreadPool.WorkerThread.cs as several new toplevel functions.

Current PR is just refactoring existing code. No functional change.

* Split PortableThreadPool.WorkerThread start and loop body

   For browser-wasm we will need to start the worker thread in a special way, and use callbacks to run the loop body.

   Current PR is just refactoring existing code. No functional change.

* Change loop to use return instead of break

* rename utility method to ShouldExitWorker