Remove unused helix queue definitions
We don't run tests on official builds since repo consolidation in .NET 5. We used to run tests on official builds back in dotnet/coreclr, but we discontinued that practice. This YAML was all included in support of doing so. As a result, it's been dead for years and just clutters up our pipeline definitions and creates confusion.