Disallow `BLK<SIMD>` (#84868)
* Add an assert

* Move gtNewStructVal/gtNewBlkIndir/gtNewIndir

From compiler.hpp to gentree.cpp.

Motivation: more methods are to come and the header file is very large already.

* Generalize "gtNewStructVal" to avoid diffs

* Stop producing BLK<SIMD>

* Fix some regressions

The IND(COMMA) -> COMMA(IND) transformation decanonicalizes array
access as it is only performed on "uses" (i. e. not stores).

Simply do not do this for SIMD types. Evidently, this unblocks some if conversions.

* Fix gtNewZeroCon

* Fix regressions related to SIMD promotion

Preserve the behavior of setting the "is used in a SIMD intrinsic" flag.

Will bring additional positive diffs due to the "ldobj" change.