DateTimeParse: accept all white space chars in the format string for DateTimeStyles Allow White
When the string is parsed using DateTimeStyles Allow White, the parser
will skip any type of white space in the parsed string. Then when the parser
tries to match with a non-regular space from the format string, it fails
because it only accepts regular spaces in the format string.

This changes the accepted white space characters for Allow White in the format
string to match with the skipped white space characters in the parsed string.