[wasm-mt] make JSHostImplementation.s_csOwnedObjects ThreadStatic
The integer jsHandles are not global - they are essentially indices
into a JS array.
So the mapping from a jsHandle to a JSObject must be per-thread.

This fixes the thread affinity assertions in the smoketest (which were
false positives - we looked up a worker's jsHandle and got back the
main thread's JSObject - and then asserted that it was accessed from
the wrong thread)