Retry compat Random.NextSingle in case rounding from double produces 1.0f
The compat implementation behind Random just casts a double to a float to produce the result of NextSingle.  Some double values close to 1.0 might round up to 1.0f, which is out of range for NextSingle.  In that case, reject it and retry.  This approach is consistent with how fairness is handled elsewhere in the implementation, and avoids changing the sequence of values returned for a fixed seed in cases where an illegal value wasn't being returned.