JIT: Generalize assignment decomposition in physical promotion
Generalize assignment decomposition to handle arbitrary combinations of
physically promoted structs. Do this by introducing a DecompositionPlan
class that keeps track of the copies to do that involve replacement
fields. The first step is then to fill out this plan. In the general
case where both the source and destination are physically promoted this
involves iterating the replacements in lockstep. For promotions that map
exactly, a direct copy between their locals is queued into the plan; in
other cases (e.g. partial overlap) it may involve writing the source
back to the struct local.

The plan is used to generate the IR and to figure out whether the writes
of all the replacement fields covers the destination such that the
original struct copy can be omitted.

Also unnest StatementList and rename it to DecompositionStatementList as
the other name conflicted with another class.